requirements/base.txt
# Django
Django==4.2.16
#pillow==10.3.0
asgiref==3.8.1

# Django 3rd party modules
django-allauth==65.0.2
django-filter==24.2
django-extensions==3.2.3
django-model-utils==4.5.1
djangorestframework==3.15.1
django-constance==4.3.2
django-easy-audit==1.3.6
django-widget-tweaks==1.5.0
django-celery-beat==2.7.0
django-sesame==3.2.2

# Environment settings
python-dotenv==1.0.1
python-decouple==3.8
dj-database-url==1.2.0

# Others
Faker==25.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
setuptools==70.0.0
coverage==7.5.3
requests==2.32.3
sentry-sdk==2.19.2

# Generate Files
python-magic==0.4.27
reportlab==4.2.2

# Pandas & Excel
pandas==2.2.2
XlsxWriter==3.2.0
openpyxl==3.1.5

# PDF
pypdf==5.0.1
PyPDF2==3.0.1
weasyprint==64.1

# Celery
celery==5.4.0
redis==5.2.1

# Test
factory_boy==3.3.3
